backend/requirements-dev.txt
# Backend Dependencies - Local Development (SQLite, Python 3.13 compatible)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
email-validator>=2.1.0
sqlalchemy>=2.0.25
aiosqlite>=0.19.0
httpx>=0.26.0
google-generativeai>=0.3.2
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
python-jose[cryptography]>=3.3.0
youtube-transcript-api>=0.6.2
python-multipart>=0.0.6
